// frontend/src/gen/types/JobOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type JobOut = {
    /**
     * @type integer
    */
    id: number;
    /**
     * @type string
    */
    shot_slug: string;
    /**
     * @type string
    */
    status: string;
    error?: (string | null);
    image_url?: (string | null);
};