// frontend/src/gen/types/StateOut.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { CharacterOut } from "./CharacterOut.ts";
import type { ContinuityOut } from "./ContinuityOut.ts";
import type { JobOut } from "./JobOut.ts";
import type { ShotOut } from "./ShotOut.ts";

export type StateOut = {
    /**
     * @default ""
     * @type string | undefined
    */
    title?: string;
    /**
     * @default ""
     * @type string | undefined
    */
    author?: string;
    /**
     * @default ""
     * @type string | undefined
    */
    provider?: string;
    /**
     * @default ""
     * @type string | undefined
    */
    style?: string;
    /**
     * @type array | undefined
    */
    characters?: CharacterOut[];
    /**
     * @type array | undefined
    */
    continuity?: ContinuityOut[];
    /**
     * @type array | undefined
    */
    jobs?: JobOut[];
    /**
     * @type array | undefined
    */
    scenes?: ShotOut[];
};